+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Create release command line utility [#6](https://github.com/stevewest/changelog/issues/6).
 - Create convert command line utility [#6](https://github.com/stevewest/changelog/issues/6).
 - Create add command line utility [#6](https://github.com/stevewest/changelog/issues/6).
+- Task to create .phar distributable.
 
 ### Fixed
 - Links are no longer parsed as a description.